+++ test-fixtures/deno-app/src/main.ts
@@ -1,1 +1,0 @@
-import { serve } from "std/http/server.ts"; serve((_req) => new Response("Hello!"));